requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
PyYAML>=6.0.1
numpy>=1.24.0
tqdm>=4.66.0
vector-quantize-pytorch>=1.14.0
opencv-python>=4.8.0
einops>=0.7.0
accelerate>=0.30.0
tensorboard>=2.12.0
scikit-learn>=1.3.0
scipy>=1.10.0
umap-learn>=0.5.5
matplotlib>=3.7.0
seaborn>=0.12.2
hdbscan>=0.8.33

# Optional: CoTracker (for real preprocessing)
# git+https://github.com/facebookresearch/co-tracker.git
